renamed joystick to gamepad as not to confuse it with sf joystick class
